Refactoring InventoryPage
