+++ SheetDice/AppShell.xaml
@@ -2,6 +2,7 @@
 <Shell xmlns="http://xamarin.com/schemas/2014/forms" 
        xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
        xmlns:local="clr-namespace:SheetDice.Views"
+       xmlns:inventario="clr-namespace:SheetDice.Views.Inventario"
        Title="SheetDice"
        x:Class="SheetDice.AppShell">
 
@@ -78,7 +79,7 @@
     <FlyoutItem FlyoutItemIsVisible="False">
         <Tab>
             <ShellContent Title="blank" Route="BlankPage" ContentTemplate="{DataTemplate local:BlankPage}" />
-            <ShellContent Title="Inventory" Route="InventoryPage" ContentTemplate="{DataTemplate local:InventoryPage}"/>
+            <ShellContent Title="Inventory" Route="InventoryPage" ContentTemplate="{DataTemplate inventario:InventoryPage}"/>
         </Tab>
     </FlyoutItem>
